Refactor specs for CarrierWave
