+++ spec/factories/movies.rb
@@ -2,6 +2,6 @@ FactoryBot.define do
   factory :movie do
     title { Faker::Movie.title }
     description { Faker::Lorem.paragraph(sentence_count: 3) }
-    image { Faker::Internet.url }
+    image { Rack::Test::UploadedFile.new(File.open(File.join(Rails.root, '/spec/fixtures/myfiles/test.jpg')), 'image/jpeg') }
   end
 end
