+++ spec/factories/users.rb
@@ -1,7 +1,7 @@
 FactoryBot.define do
   factory :user do
     name { Faker::Internet.username(specifier: 3..12) }
-    admin_level { 0 }
+    admin_level { 3 }
     password { Faker::Internet.password }
     email { Faker::Internet.email }
   end
