+++ spec/models/movie_spec.rb
@@ -14,6 +14,4 @@ RSpec.describe Movie, type: :model do
   it { should validate_length_of(:description).is_at_least(10).is_at_most(255) }
 
   it { should validate_presence_of(:image) }
-  it { should allow_value('https://foo.com').for(:image) }
-  it { should_not allow_value('foo.bar').for(:image) }
 end
