+++ spec/requests/movies_request_spec.rb
@@ -3,7 +3,8 @@ require 'rails_helper'
 RSpec.describe 'Movies', type: :request do
   let!(:movie) { create(:movie) }
   let(:movie_id) { movie.id }
-
+  let(:user) { create(:user) }
+ 
   describe 'GET /movies' do
     before { get '/movies' }
 
@@ -38,10 +39,12 @@ RSpec.describe 'Movies', type: :request do
   end
 
   describe 'POST /movies' do
-    let(:valid_attributes) { movie.attributes.merge(title: Faker::Movie.title, categories: %w[Action Adventure]) }
-
+    file = Rack::Test::UploadedFile.new(File.open(File.join(Rails.root, '/spec/fixtures/myfiles/test.jpg')), 'image/jpeg')
+    let(:valid_attributes) { movie.attributes.merge(title: Faker::Movie.title, categories: %w[Action Adventure], image: file) }
+    
     context 'when the request is valid' do
       before do
+        set_session(user_id: user.id)
         Category.create(name: 'Action')
         Category.create(name: 'Adventure')
         post '/movies', params: { movie: valid_attributes }
@@ -63,6 +66,7 @@ RSpec.describe 'Movies', type: :request do
 
     context 'when the request is invalid' do
       before do
+        set_session(user_id: user.id)
         post '/movies', params: { movie: { title: 'Green Mile', categories: [] } }
       end
 
@@ -77,6 +81,7 @@ RSpec.describe 'Movies', type: :request do
 
     context 'when the record exists' do
       before do
+        set_session(user_id: user.id)
         Category.create(name: 'Action')
         put "/movies/#{movie_id}", params: { movie: valid_attributes }
       end
@@ -92,7 +97,10 @@ RSpec.describe 'Movies', type: :request do
   end
 
   describe 'DELETE /movies/:id' do
-    before { delete "/movies/#{movie_id}" }
+    before do 
+      set_session(user_id: user.id)
+      delete "/movies/#{movie_id}"
+    end
 
     it 'deletes record' do
       expect(Movie.find_by(id: movie_id)).to be_nil
